refactor: remove redundant files, update docker and server logic

+++ download-piper.sh
@@ -1,36 +1,44 @@
-#!/bin/bash
+#!/bin/sh
 
-# Function to get the machine architecture
 get_architecture() {
     case $(uname -m) in
-        x86_64) echo "amd64" ;;
-        aarch64) echo "arm64" ;;
+        x86_64 | amd64) echo "x86_64" ;;
+        aarch64 | arm64) echo "aarch64" ;;
         armv7l) echo "armv7" ;;
+    esac
+}
+
+get_os() {
+    case $(uname -s) in
+        Linux) echo "linux" ;;
+        Darwin) echo "macos" ;;
     esac
 }
 
 # Get the machine architecture
 architecture=$(get_architecture)
+if [ -z "$architecture" ]; then
+    echo "Unsupported architecture"
+    exit 1
+fi
 
-# GitHub URL
-url="https://github.com/rhasspy/piper/releases/latest"
-
-# Get the redirect URL of the latest release
-redirect_url=$(curl -sL -w %{url_effective} -o /dev/null $url)
-
-# Extract the release tag from the redirect URL
-release_tag=$(basename $redirect_url)
-
-# Construct the release file URL
-release_file_url="https://github.com/rhasspy/piper/releases/download/$release_tag/piper_$architecture.tar.gz"
+# Get the OS
+os=$(get_os)
+if [ -z "$os" ]; then
+    echo "Unsupported OS"
+    exit 1
+fi
 
 # Download the release file
-curl -L -o piper.tar.gz $release_file_url
+url="https://github.com/rhasspy/piper/releases/latest/download/piper_${os}_${architecture}.tar.gz"
+curl -L -o piper.tar.gz "$url"
 
-mkdir bin
+# Create the bin directory if it doesn't exist or clean it up
+mkdir -p ./bin
+rm -rf ./bin/*
 
 # Extract the tar gz archive
-tar -xzf piper.tar.gz -C bin
+tar -xzf piper.tar.gz -C ./bin
 
 # Clean up the downloaded archive
 rm piper.tar.gz
